Penambahan Tampilan Dashboard

+++ app/Http/Controllers/KeluargaController.php
@@ -34,6 +34,7 @@ class KeluargaController extends Controller
             'foto' => 'nullable|image|mimes:jpg,jpeg,webp,png|max:5048',
             'tanggal_lahir' => 'required|date',
             'alamat' => 'required|string',
+            'link_gmap' => 'nullable|string|max:1000',
             'status_keluarga_id' => 'required|exists:status_keluargas,id',
             'anak_ke' => 'required|integer',
             'keluarga_id' => 'nullable|exists:keluargas,id',
@@ -52,6 +53,7 @@ class KeluargaController extends Controller
             'foto' => $imagePath,
             'tanggal_lahir' => $request->tanggal_lahir,
             'alamat' => $request->alamat,
+            'link_gmap' => $request->link_gmap,
             'status_keluarga_id' => $request->status_keluarga_id,
             'anak_ke' => $request->anak_ke,
             'keluarga_id' => $request->keluarga_id,
@@ -70,41 +72,46 @@ class KeluargaController extends Controller
 
     public function edit(Keluarga $keluarga)
     {
-        $statuses = StatusKeluarga::all(); // Untuk dropdown Status Keluarga
-        $keluargas = Keluarga::all(); // Untuk dropdown Orang Tua dan Pasangan
+        $statuses = StatusKeluarga::all();
+        $keluargas = Keluarga::all();
         return view('keluarga.edit', compact('keluarga', 'statuses', 'keluargas'));
     }
 
     public function update(Request $request, Keluarga $keluarga)
     {
-        // Validasi input
         $request->validate([
             'nama' => 'required|string|max:255',
-            'foto' => 'nullable|image|mimes:jpg,jpeg,png,gif|max:5048', // Validasi foto
+            'foto' => 'nullable|image|mimes:jpg,jpeg,png,gif|max:5048',
             'tanggal_lahir' => 'required|date',
             'alamat' => 'required|string',
+            'link_gmap' => 'nullable|string|max:1000',
             'status_keluarga_id' => 'required|exists:status_keluargas,id',
             'anak_ke' => 'required|integer',
-            'keluarga_id' => 'nullable|exists:keluargas,id', // Validasi keluarga (orang tua)
+            'keluarga_id' => 'nullable|exists:keluargas,id',
             'pasangan_id' => 'nullable|exists:keluargas,id|different:keluarga_id',
             'status' => 'required|in:hidup,meninggal',
             'tanggal_meninggal' => 'nullable|date',
         ]);
 
-        // Jika ada foto baru, hapus foto lama dan simpan foto baru
+        if ($request->hapus_foto == "1" && $keluarga->foto) {
+            Storage::disk('public')->delete($keluarga->foto);
+            $keluarga->foto = null;
+        }
+
         if ($request->hasFile('foto')) {
             if ($keluarga->foto) {
-                Storage::disk('public')->delete($keluarga->foto); // Menghapus foto lama
+                Storage::disk('public')->delete($keluarga->foto);
             }
             $fotoPath = $request->file('foto')->store('keluarga', 'public');
             $keluarga->foto = $fotoPath;
         }
 
-        // Update data keluarga
         $keluarga->update([
             'nama' => $request->nama,
+            'foto' => $keluarga->foto,
             'tanggal_lahir' => $request->tanggal_lahir,
             'alamat' => $request->alamat,
+            'link_gmap' => $request->link_gmap,
             'status_keluarga_id' => $request->status_keluarga_id,
             'anak_ke' => $request->anak_ke,
             'keluarga_id' => $request->keluarga_id,
@@ -113,13 +120,13 @@ class KeluargaController extends Controller
             'tanggal_meninggal' => $request->tanggal_meninggal,
         ]);
 
-        // return redirect()->route('keluarga.index')->with('success', 'Keluarga berhasil diperbarui.');
         if (Auth::check()) {
-            return redirect()->route('keluarga.index')->with('success', 'Keluarga berhasil ditambahkan');
+            return redirect()->route('keluarga.index')->with('success', 'Keluarga berhasil diperbarui.');
         } else {
             return redirect('/silsilah')->with('warning', 'Silakan login untuk mengakses fitur ini.');
         }
     }
+
 
     public function destroy(Keluarga $keluarga)
     {
